Do not support stream type "b"
Each stream has unique ID, so in case if configuration option
"both" is provided there should be 2 IDs, but there is a
single xenstore entry for this. If we allocate in FE 2 IDs,
with the one read from the strore being base number,
then this could be implementation dependant: which one is
for playback and which for capture.
Make things clearer and define streams implicitly.

Signed-off-by: Andrushchenko, Oleksandr <andr2000@gmail.com>